Allow whitespace around version in dependency specification.
Previously the parsing of a FW_PACKAGE_DEPENDS string such as
"package ( < 1.2.3 )" would silently ignore the version part because of
the whitespace after the opening "(" and before the closing ")" of the
version dependency.  The regular expression that parses the dependency
strings now allows whitespace around the parens of a version
specification and the square brackets of an architecture specification.