FwRpmPackage: Do loose enforcement of release equality when verifying dependencies.
An installed RPM may have a dependency like "tcl = 8.5.7" that is
satisfied by an installed RPM tcl-8.5.7-6.  Previously rpmevrcmp would
consider this equality dependency to be violated.  This revision changes
rpmevrcmp so that when its second argument, which is the dependency
version, does not contain a release part, than any release in the
installed version is considered equal.