rpm/make-package: Don't exclude versioned directories from RPM ownership.
make-package excludes a directory from the %files list if it already
exists on the host.  Unfortunately, that means that if you are
rebuilding an RPM for an package that is already installed the RPM will
not end up owning its directories.

This change is not the best fix, but it fixes the specific problem we're
having with rebuilding versioned Erlang packages.